# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Rasan SaaS Chatbot Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and receive auth cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and clear auth cookie
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Tenants ──────────────────────────────────────────────────────────────────
...
  /super-admin/tenants:
    get:
      operationId: listAllTenants
      tags: [super-admin]
      summary: List all tenants
      security:
        - cookieAuth: []
      responses:
        "200":
          description: All tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantSummary"

    post:
      operationId: createTenant
      tags: [super-admin]
      summary: Create a new tenant
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantSummary"
  /super-admin/tenants/{tenantId}:
    patch:
      operationId: updateTenantAdmin
      tags: [super-admin]
      summary: Update tenant (admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantAdminRequest"
      responses:
        "200":
          description: Updated tenant

    delete:
      operationId: deleteTenant
      tags: [super-admin]
      summary: Delete a tenant
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      responses:
        "204":
          description: Tenant deleted
  /super-admin/tenants/{tenantId}/users:
    get:
      operationId: listTenantUsers
      tags: [super-admin]
      summary: List users for a tenant
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantUser"

    post:
      operationId: createTenantUser
      tags: [super-admin]
      summary: Create a user for a tenant
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
  /super-admin/users/{userId}:
    patch:
      operationId: updateTenantUser
      tags: [super-admin]
      summary: Update a tenant user
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user

    delete:
      operationId: deleteTenantUser
      tags: [super-admin]
      summary: Delete a tenant user
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deleted
  /super-admin/settings:
    get:
      operationId: getSystemSettings
      tags: [super-admin]
      summary: Get system settings
      security:
        - cookieAuth: []
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string

    put:
      operationId: updateSystemSettings
      tags: [super-admin]
      summary: Update system settings
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                openai_api_key:
                  type: string
      responses:
        "200":
          description: Settings updated

# ─── Components ───────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rasan_token
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
      required: [detail]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        superAdmin:
          type: boolean
    AuthMeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        tenant_id:
          type: string
          format: uuid
        subscription_tier:
          type: string
          enum: [free, starter, pro, business]
      required: [id, email, role]
...
    TenantSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        subscriptionTier:
          type: string
        isActive:
          type: boolean
        messagesUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, subscriptionTier, isActive]
...
    UpdateTenantAdminRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subscriptionTier:
          type: string
          enum: [free, starter, pro, business]
        isActive:
          type: boolean
    CreateTenantRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subscriptionTier:
          type: string
          enum: [free, starter, pro, business]
        isActive:
          type: boolean
      required: [name]
...
    TenantUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, agent]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, role, isActive]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, agent]
      required: [email, password]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, agent]
        isActive:
          type: boolean
        password:
          type: string